chore: Minor update
### Added
- Added font8x8 from https://github.com/dhepper/font8x8
- Added WIP SDL2 lib implementation
- Added file read and write implementation
### Changed
- Change static lib name from `jadeitite` to `jadeitite_glut`
- Add simple drawing for example
- Change example name for `base_template` to `glut_template`
### Fixed
- Setup macro for system config for `TCC` compiler

+++ src/jadeitite/renderer/gl.c
@@ -28,9 +28,9 @@ void draw_point_int(int p_posX, int p_posY) {
  *
  * @warning Be aware of pointers and count, there are no checking and can crash on overflow!
  */
-void draw_points_int(int *p_posX, int *p_posY, size_t p_count) {
+void draw_points_int(int *p_posX, int *p_posY, u32 p_count) {
   glBegin(GL_POINTS);
-  for (size_t i = 0; i < p_count; ++i) {
+  for (u32 i = 0; i < p_count; ++i) {
     glVertex2i(p_posX[i], p_posY[i]);
   }
   glEnd();
